lib:serial_posix: use lsof to check port
Check tty port prior to using it. use lsof
command to check open processes. lsof is more of
a standard *ix app is available almost all distributions

The only other alternative would be to write a logic to
parse through the pids in the system looking for user of
the device node :(..

using grep based on feedback from robclark for mac OS X

Signed-off-by: Nishanth Menon <nm@ti.com>