Fixed bug where binaries will not run on PowerPC.
The main issue was with the way getopt from libc6 is used. It seems like
the way glib resets getopt is different. More information about this
issue can be found in

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=192834

Looking at the example provided in Message#53, the return for getopt is
an int. Changing from char to int in the files fixes the issue.

	modified:   src/gpsign.c
	modified:   src/pserial.c
	modified:   src/ucmd.c
	modified:   src/ukermit.c
Signed-off-by: Amit Uttamchandani <amit.uttam@gmail.com>
Signed-off-by: Nishanth Menon <nm@ti.com>