tagger: a tag header generation utility
Arm linux kernel has a bunch of boot requirements that are listed:
http://www.arm.linux.org.uk/developer/booting.php

tagger is an utility used to generate this tag along with a small
sty - an assembly util which gets control when system starts off
the sty puts the data in required locations and gives control
to the kernel's piggy allowing for the system to bootup

when used in conjunction with gpsign, this could replace the
traditional bootloader in some cases.

Benefit ranges from:
* quick boot of the kernel
* removing dependency of kernel on a specific bootloader (e.g. mux etc..)

Example usage:
in the kernel: make bzImage

./tagger -c configs/sample-3630.cfg -f zImage
./gpsign -c configs/sample-3630.cfg -f zImage.tag

tagger generates an output file .tag which can be used along with
u-boot "go" command or equivalents such as csst or other flashing
utilities designed to execute images from SDRAM

gpsign adds a configuration header and generates a .ift which
is a flashable image which is downloaded to a I/O image at 0
e.g. NAND flash (remember this should be written with hw ecc to NAND)

Signed-off-by: Nishanth Menon <nm@ti.com>