pusb: cannot support GET_ASICID on OMAP3
commit d086cb5c0393306d35177b1c024a8fa7568e1329
(pusb: add support for omap 4430)

introduced support for using GET_ASICID, however,
3430 has no clue about what GET_ASICID is. So, be smart
and send command only if length is not what we expect
it to be. This should continue to support both OMAP3 and 4.

Signed-off-by: Pali Rohár <pali.rohar@gmail.com>
[nm@ti.com: modified original patch - the original patch is available:
https://groups.google.com/forum/?fromgroups=#!topic/beagleboard/bj8qRG82ssY
However, the fix need not need have option provided from commandline
OMAP3 ROMCode is just NOT smart enough to figure out about GET_ASICID-
you either get the data OR not.]
Signed-off-by: Nishanth Menon <menon.nishanth@gmail.com>