pusb: convert to libusb-1.0
libusb-1.0 seems to be the most maintained library version, so
first convert to that library which has a different api.
libusb-1.0 has the nice function libusb_open_device_with_vid_pid()
which does everything needed to open that particular device, so no
need to scan the bus manually, it would be only interesting for
debugging purposes and lsusb -v gives you also the same output as
you would get by bus scanning. So for simplicity the bus scan was
not translated to the newer api.

Signed-off-by: Andreas Kemnade <andreas@kemnade.info>